Move code generators into Core and make AppCapabilities more general
