+++ Core/CMakeLists.txt
@@ -43,7 +43,10 @@ set(PUBLIC_HEADERS
     cafApplication.h
     cafSession.h
     cafUuidGenerator.h
-    cafFieldDocumentationCapability.h)
+    cafFieldDocumentationCapability.h
+    cafCodeGenerator.h
+    cafMarkdownGenerator.h
+    cafPythonGenerator.h)
 
 set(PROJECT_FILES
     ${PUBLIC_HEADERS}
@@ -65,7 +68,9 @@ set(PROJECT_FILES
     cafApplication.cpp
     cafSession.cpp
     cafUuidGenerator.cpp
-    cafFieldDocumentationCapability.cpp)
+    cafFieldDocumentationCapability.cpp
+    cafMarkdownGenerator.cpp
+    cafPythonGenerator.cpp)
 
 if(CAFFA_BUILD_SHARED)
   message(STATUS "Building ${PROJECT_NAME} shared")
