+++ Core/cafApplication.h
@@ -43,11 +43,10 @@ struct AppInfo
      */
     enum class AppCapability : unsigned int
     {
-        CONSOLE     = 0x00, ///< Console Application
-        GRPC_SERVER = 0x01, ///< gRPC server
-        GRPC_CLIENT = 0x02, ///< gRPC client
-        GUI         = 0x04, ///< GUI application
-        WEB         = 0x08 ///< Web server
+        CONSOLE = 0x00, ///< Console Application
+        SERVER  = 0x01, ///< Server
+        CLIENT  = 0x02, ///< Client
+        GUI     = 0x04 ///< GUI application
     };
 
     /**
@@ -71,7 +70,7 @@ struct AppInfo
      */
     int patchVersion;
     /**
-     * @brief Application type. Can be CONSOLE, GRPC_SERVER, GRPC_CLIENT, GUI or WEB.
+     * @brief Application type. Can be CONSOLE, SERVER, CLIENT, GUI
      *
      */
     unsigned int appType;
