+++ GrpcInterface/cafGrpcAppService.cpp
@@ -63,7 +63,7 @@ grpc::Status AppService::PerformGetAppInfo( grpc::ServerContext* context, const 
     CAFFA_TRACE( "Received app info request from " + context->peer() );
     Application* app = Application::instance();
     reply->set_name( app->name() );
-    CAFFA_ASSERT( app->hasCapability( AppInfo::AppCapability::GRPC_SERVER ) );
+    CAFFA_ASSERT( app->hasCapability( AppInfo::AppCapability::SERVER ) );
 
     AppInfo appInfo = app->appInfo();
     reply->set_type( appInfo.appType );
