+++ GrpcInterface/cafGrpcServerApplication.cpp
@@ -1,20 +1,20 @@
-//##################################################################################################
+// ##################################################################################################
 //
-//   Caffa
-//   Copyright (C) 2021- 3D-Radar AS
+//    Caffa
+//    Copyright (C) 2021- Kontur AS
 //
-//   GNU Lesser General Public License Usage
-//   This library is free software; you can redistribute it and/or modify
-//   it under the terms of the GNU Lesser General Public License as published by
-//   the Free Software Foundation; either version 2.1 of the License, or
-//   (at your option) any later version.
+//    GNU Lesser General Public License Usage
+//    This library is free software; you can redistribute it and/or modify
+//    it under the terms of the GNU Lesser General Public License as published by
+//    the Free Software Foundation; either version 2.1 of the License, or
+//    (at your option) any later version.
 //
-//   This library is distributed in the hope that it will be useful, but WITHOUT ANY
-//   WARRANTY; without even the implied warranty of MERCHANTABILITY or
-//   FITNESS FOR A PARTICULAR PURPOSE.
+//    This library is distributed in the hope that it will be useful, but WITHOUT ANY
+//    WARRANTY; without even the implied warranty of MERCHANTABILITY or
+//    FITNESS FOR A PARTICULAR PURPOSE.
 //
-//   See the GNU Lesser General Public License at <<http://www.gnu.org/licenses/lgpl-2.1.html>>
-//   for more details.
+//    See the GNU Lesser General Public License at <<http://www.gnu.org/licenses/lgpl-2.1.html>>
+//    for more details.
 //
 #include "cafGrpcServerApplication.h"
 
@@ -35,7 +35,7 @@ ServerApplication::ServerApplication( int                portNumber,
                                       const std::string& serverCertFile /* = ""*/,
                                       const std::string& serverKeyFile /* = ""*/,
                                       const std::string& caCertFile /* = ""*/ )
-    : Application( AppInfo::AppCapability::GRPC_SERVER )
+    : Application( AppInfo::AppCapability::SERVER )
 
 {
     m_server = std::make_unique<caffa::rpc::Server>( portNumber, serverCertFile, serverKeyFile, caCertFile );
